Implementing generic handler
