+++ src/main/java/com/luminary/apieden/controller/UserController.java
@@ -93,23 +93,4 @@ public class UserController implements UserContract {
         userService.deleteById(id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
-
-
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public Map<String, Object> verifyUser(BindingResult result) {
-        log.error("BAD REQUEST error. {}", result);
-        Map<String, Object> errors = new HashMap<>();
-        for (FieldError error: result.getFieldErrors()) {
-            errors.put(error.getField(), error.getDefaultMessage());
-        }
-        return errors;
-    }
-
-    @ExceptionHandler(HttpError.class)
-    public ResponseEntity<ErrorResponse> genericHandler(HttpError error) {
-        log.error("An error occurred. {}", error.getMessage());
-        ErrorResponse errorResponse = new ErrorResponse(error.getStatus(), error.getMessage());
-        return ResponseEntity.status(errorResponse.getHttpStatus()).body(errorResponse);
-    }
 }
