+++ src/main/java/com/luminary/apieden/model/exception/HttpError.java
@@ -7,6 +7,6 @@ import org.springframework.http.HttpStatus;
 @Getter
 @AllArgsConstructor
 public class HttpError extends RuntimeException{
-    private final HttpStatus status;
+    private final HttpStatus httpStatus;
     private final String message;
 }
